Fixed issues where rapidly deleting/adding items would cause concurrency issues while saving new state to disk.
Technical details on the issue: the user remove an item. The ObservableList would get passed to gson for serialization, but while gson was iterating, the user was able to remove another item, thus breaking the contract/iteration for the gson iteration. My solution was that everytime the service which saves state to disk is fired, a deep copy is passed to it, so it can operate freely on it. if another deletion is made, the first service is first canceled then restarted, so at most we have 2x the memory consumed by the list. This is not ideal but allow for concurrent access to the list.

Closes #18